+++ src/Core/Entity/EntityLooper.js
